+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/component/EntityComboBoxPanel.java
@@ -242,7 +242,7 @@ public final class EntityComboBoxPanel extends JPanel {
 		}
 
 		private EntityComboBox createComboBox() {
-			return entityComboBoxBuilder.clear().build();
+			return entityComboBoxBuilder.build();
 		}
 
 		private static class EntityComboBoxPanelValue extends AbstractComponentValue<Entity, EntityComboBoxPanel> {